Moved bestiary to the correct directory
